feat: mig-smoke-ready-deletion-batches-008 - Migrate smoke-delete-04 transport CLI
