+++ tests/functional/smoke/packaged_goal_parity_helpers_test.go
@@ -1,0 +1,175 @@
+package smoke
+
+import (
+	"encoding/json"
+	"fmt"
+	"net"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/portpowered/infinite-you/internal/testutil"
+	modelprovider "github.com/portpowered/infinite-you/pkg/services/models"
+	"github.com/portpowered/infinite-you/pkg/services/workers"
+	"github.com/portpowered/infinite-you/tests/functional/internal/support"
+)
+
+const defaultPromptRunWorkTypeName = "prompt-task"
+
+func scaffoldPackagedGoalInvocationFactoryForSmoke(t *testing.T) string {
+	t.Helper()
+
+	cfg := factoryPromptRunSmokeConfig()
+	cfg["name"] = "@you/goal"
+	cfg["invocationReturn"] = map[string]any{
+		"policy":        "EXPLICIT",
+		"workTypeName":  "goal",
+		"terminalState": "complete",
+	}
+	workTypes := cfg["workTypes"].([]map[string]any)
+	workTypes[0]["name"] = "goal"
+	workstations := cfg["workstations"].([]map[string]any)
+	workstations[0]["name"] = "execute-goal"
+	workstations[0]["worker"] = "goal-executor"
+	for _, ioKey := range []string{"inputs", "outputs", "onFailure"} {
+		ios := workstations[0][ioKey].([]map[string]string)
+		for i := range ios {
+			ios[i]["workType"] = "goal"
+		}
+	}
+	cfg["workers"] = []map[string]string{{"name": "goal-executor"}}
+
+	dir := support.ScaffoldFactory(t, cfg)
+	support.WriteAgentConfig(
+		t,
+		dir,
+		"goal-executor",
+		support.BuildModelWorkerConfig(modelprovider.ProviderCodex, "gpt-5-codex"),
+	)
+	return dir
+}
+
+func factoryPromptRunSmokeConfig() map[string]any {
+	return map[string]any{
+		"name": "factory-prompt-run-smoke",
+		"workTypes": []map[string]any{
+			{
+				"name":             defaultPromptRunWorkTypeName,
+				"handlingBehavior": []string{"DEFAULT"},
+				"states":           promptRunWorkTypeStates(),
+			},
+		},
+		"workers": []map[string]string{
+			{"name": "mock-worker"},
+		},
+		"workstations": []map[string]any{
+			{
+				"name":      "process-prompt",
+				"worker":    "mock-worker",
+				"inputs":    []map[string]string{{"workType": defaultPromptRunWorkTypeName, "state": "init"}},
+				"outputs":   []map[string]string{{"workType": defaultPromptRunWorkTypeName, "state": "complete"}},
+				"onFailure": []map[string]string{{"workType": defaultPromptRunWorkTypeName, "state": "failed"}},
+			},
+		},
+	}
+}
+
+func promptRunWorkTypeStates() []map[string]string {
+	return []map[string]string{
+		{"name": "init", "type": "INITIAL"},
+		{"name": "complete", "type": "TERMINAL"},
+		{"name": "failed", "type": "FAILED"},
+	}
+}
+
+func writeDefaultMockWorkersConfig(t *testing.T) string {
+	t.Helper()
+
+	data, err := json.MarshalIndent(workers.NewEmptyMockWorkersConfig(), "", "  ")
+	if err != nil {
+		t.Fatalf("marshal default mock-workers config: %v", err)
+	}
+	path := filepath.Join(t.TempDir(), "mock-workers.json")
+	if err := os.WriteFile(path, data, 0o644); err != nil {
+		t.Fatalf("write mock-workers config: %v", err)
+	}
+	return path
+}
+
+func writePackagedGoalBuiltinMockWorkersConfig(t *testing.T) string {
+	t.Helper()
+
+	cfg := workers.MockWorkersConfig{
+		MockWorkers: []workers.MockWorkerConfig{
+			{
+				WorkerName:      "goal-planner",
+				WorkstationName: "plan-goal",
+				RunType:         workers.MockWorkerRunTypeAccept,
+			},
+			{
+				WorkerName:      "goal-executor",
+				WorkstationName: "execute-goal",
+				RunType:         workers.MockWorkerRunTypeAccept,
+			},
+			{
+				WorkerName:      "goal-checker",
+				WorkstationName: "check-goal",
+				RunType:         workers.MockWorkerRunTypeScript,
+				ScriptConfig: &workers.MockWorkerScriptConfig{
+					Command: "/bin/echo",
+					Args:    []string{"plain"},
+				},
+			},
+			{
+				WorkerName:      "goal-reviewer",
+				WorkstationName: "review-goal",
+				RunType:         workers.MockWorkerRunTypeScript,
+				ScriptConfig: &workers.MockWorkerScriptConfig{
+					Command: "/bin/echo",
+					Args:    []string{"accepted"},
+				},
+			},
+		},
+	}
+	data, err := json.MarshalIndent(cfg, "", "  ")
+	if err != nil {
+		t.Fatalf("marshal packaged goal mock-workers config: %v", err)
+	}
+	path := filepath.Join(t.TempDir(), "mock-workers-packaged-goal.json")
+	if err := os.WriteFile(path, data, 0o644); err != nil {
+		t.Fatalf("write packaged goal mock-workers config: %v", err)
+	}
+	return path
+}
+
+func buildYouCLIBinary(t *testing.T) string {
+	t.Helper()
+
+	binaryName := "you"
+	if runtime.GOOS == "windows" {
+		binaryName += ".exe"
+	}
+	binaryPath := filepath.Join(t.TempDir(), binaryName)
+	build := exec.Command("go", "build", "-o", binaryPath, "./cmd/factory")
+	build.Dir = testutil.MustRepoRoot(t)
+	if output, err := build.CombinedOutput(); err != nil {
+		t.Fatalf("build you CLI: %v\n%s", err, string(output))
+	}
+	return binaryPath
+}
+
+func reserveLocalTCPPort() (int, error) {
+	listener, err := net.Listen("tcp", "127.0.0.1:0")
+	if err != nil {
+		return 0, err
+	}
+	defer listener.Close()
+
+	addr, ok := listener.Addr().(*net.TCPAddr)
+	if !ok {
+		return 0, fmt.Errorf("unexpected listener address type %T", listener.Addr())
+	}
+	return addr.Port, nil
+}
